Allow specifying crypto provider (ring vs aws-lc-rs).
Since aws-lc-rs gains traction, many libraries depend on it
unconditionally. In case when we have such transitive dependency, it
is hard to opt-out from either. That forces us to chose rustls' crypto
provider explicitely in code, which is not optimal.

This change keeps ring as a default provider, allowing us to opt-out
with `--no-default-features`, and opt-in to `aws-lc-rs` with
`--features aws-lc-rs`.